Code/requirements.txt
wikipedia-api==0.6.0
python-dotenv
pandas
requests
openai
anthropic
google-genai
scikit-learn
numpy
openpyxl
tiktoken
matplotlib
marimo
plotly
